individual locks

+++ src/seer/automation/autofix/agent_context.py
@@ -5,10 +5,9 @@ import os
 import shutil
 import tarfile
 import tempfile
-import time
+from random import randint
 from typing import List
 
-import joblib
 import requests
 import sentry_sdk
 import torch
@@ -82,7 +81,8 @@ class AgentContext:
 
         if tmp_dir is None:
             tmp_dir = os.path.join(
-                tempfile.gettempdir(), f"{repo_owner}-{repo_name}_{self.base_sha}"
+                tempfile.gettempdir(),
+                f"{repo_owner}-{repo_name}_{self.base_sha}_{randint(0, 100000)}",
             )
         self.tmp_dir = tmp_dir
         self.tmp_repo_path = os.path.join(self.tmp_dir, f"repo")
@@ -229,44 +229,51 @@ class AgentContext:
         return changed_files, removed_files
 
     def _get_data(self):
-        documents, nodes = self._load_data_from_github()
+        # Because we acquire separate read and write locks with the embedding step in between; there is a chance that
+        # we may be embedding the same diff twice. But this should be a rare case and otherwise we would be blocking multiple
+        # requests from running in parallel if we acquire a lock for the entire process of reading and writing the index.
+        try:
+            documents, nodes = self._load_data_from_github()
 
-        logger.debug(f"Loading index from storage context")
+            logger.debug(f"Loading index from storage context")
 
-        storage_path = os.path.join("./", "models/autofix_storage_context/")
+            storage_dir = os.path.join("./", "models/autofix_storage_context/")
+            lock_path = os.path.join(storage_dir, "lock")
 
-        with SoftFileLock(
-            os.path.join(storage_path, "lock"), timeout=60 * 60
-        ):  # 1 hour max lock timeout
-            logger.debug(f"Acquired lock for {storage_path}")
             with sentry_sdk.start_span(
                 op="seer.automation.autofix.index_loading",
                 description="Loading the vector store index from local filesystem",
             ) as span:
-                with open(self.cached_commit_json_path, "r") as file:
-                    cached_commit_data = json.load(file)
-                cached_commit_sha = cached_commit_data.get("sha")
+                with SoftFileLock(
+                    os.path.join(lock_path), timeout=60 * 60
+                ):  # 1 hour max lock timeout
+                    logger.debug(f"Acquired read lock for {storage_dir}")
+                    with open(self.cached_commit_json_path, "r") as file:
+                        cached_commit_data = json.load(file)
+                    cached_commit_sha = cached_commit_data.get("sha")
 
-                service_context = ServiceContext.from_defaults(embed_model=self.embed_model)
-                memory_vector_store = MemoryVectorStore().from_persist_dir(storage_path)
-                storage_context = StorageContext.from_defaults(
-                    vector_store=memory_vector_store,
-                    persist_dir=storage_path,
-                )
-                index_structs = storage_context.index_store.index_structs()
+                    service_context = ServiceContext.from_defaults(embed_model=self.embed_model)
+                    memory_vector_store = MemoryVectorStore().from_persist_dir(storage_dir)
+                    storage_context = StorageContext.from_defaults(
+                        vector_store=memory_vector_store,
+                        persist_dir=storage_dir,
+                    )
+                    index_structs = storage_context.index_store.index_structs()
 
-                if len(index_structs) == 0:
-                    raise Exception("No index structures found in storage context")
-                index_struct: IndexDict = index_structs[0]  # type: ignore
+                    if len(index_structs) == 0:
+                        raise Exception("No index structures found in storage context")
+                    index_struct: IndexDict = index_structs[0]  # type: ignore
 
-                index = VectorStoreIndex(
-                    index_struct=index_struct,
-                    service_context=service_context,
-                    storage_context=storage_context,
-                    show_progress=True,
-                )
+                    index = VectorStoreIndex(
+                        index_struct=index_struct,
+                        service_context=service_context,
+                        storage_context=storage_context,
+                        show_progress=True,
+                    )
 
-            logger.debug(f"Loaded index from storage context '{storage_path}'.")
+                    logger.debug(f"Released read lock for {storage_dir}")
+
+            logger.debug(f"Loaded index from storage context '{storage_dir}'.")
 
             # Update the documents that changed in the diff.
             changed_files, deleted_files = self._get_commit_file_diffs(cached_commit_sha)
@@ -307,20 +314,28 @@ class AgentContext:
 
             if commit_is_newer:
                 # Save only when the commit is newer
-                storage_context.persist(persist_dir=storage_path)
-                memory_vector_store.persist(
-                    persist_path=os.path.join(storage_path, "vector_store.json")
-                )
+                with SoftFileLock(
+                    os.path.join(lock_path), timeout=60 * 60
+                ):  # 1 hour max lock timeout
+                    logger.debug(f"Acquired write lock for {storage_dir}")
+                    storage_context.persist(persist_dir=storage_dir)
+                    memory_vector_store.persist(
+                        persist_path=os.path.join(storage_dir, "vector_store.json")
+                    )
 
-                try:
-                    os.remove(os.path.join(storage_path, "default__vector_store.json"))
-                except OSError as e:
-                    logger.error(f"Failed to delete default vector store file: {e}")
+                    try:
+                        os.remove(os.path.join(storage_dir, "default__vector_store.json"))
+                    except OSError as e:
+                        logger.error(f"Failed to delete default vector store file: {e}")
 
-                with open(self.cached_commit_json_path, "w") as sha_file:
-                    json.dump({"sha": self.base_sha}, sha_file)
+                    with open(self.cached_commit_json_path, "w") as sha_file:
+                        json.dump({"sha": self.base_sha}, sha_file)
 
-            logger.debug(f"Released lock for {storage_path}")
+                    logger.debug(f"Released write lock for {storage_dir}")
+
+        finally:
+            # Always cleanup the tmp dir
+            self.cleanup()
 
     def _get_document(self, file_path: str):
         for document in self.documents:
@@ -360,8 +375,11 @@ class AgentContext:
                 self.nodes.extend(new_nodes)
 
     def cleanup(self):
-        shutil.rmtree(self.tmp_dir)
-        logger.info(f"Deleted tmp dir {self.tmp_dir}")
+        if os.path.exists(self.tmp_dir):
+            shutil.rmtree(self.tmp_dir)
+            logger.info(f"Deleted tmp dir {self.tmp_dir}")
+        else:
+            logger.info(f"Tmp dir {self.tmp_dir} already removed")
 
     def _create_branch(self, branch_name, base_branch=None, base_commit_sha: str | None = None):
         base_sha = base_commit_sha
